Fix date.c build in msvc.
Ported the win32 implementations of gmtime_r,
localtime_r, and gettimeofday to be part of the 
posix compatibility layer, and fixed
git_signature_now to use them.

+++ src/util.h
@@ -213,7 +213,7 @@ GIT_INLINE(int) git__time_cmp(const git_time *a, const git_time *b)
 {
    /* Adjust for time zones. Times are in seconds, offsets are in minutes. */
    git_time_t adjusted_a = a->time + ((b->offset - a->offset) * 60);
-   return adjusted_a - b->time;
+   return (int)(adjusted_a - b->time);
 }
 
 #endif /* INCLUDE_util_h__ */
